Extract LocalNavHashLink
Starting to pull out the sidenav into its own component so we can stop
using the Explore css everywhere
